refactor: move prisma.service to its own folder
