+++ src/shared/shared.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { PrismaService } from './services/prisma.service';
+import { PrismaService } from './services/prisma/prisma.service';
 import { LoggingService } from './services/logging/logging.service';
 import { LogWriterService } from './services/log-writer/log-writer.service';
